Fixed closed issue
by clicking closed, the status of the issue was not changing to closed. There were two problems.

Firstly, the name of the function being called was wrong in the closed buttons, and therefore changed it.

Secondly, the id being checked inside the function with the id of the issue returned by from the localStorage had a type problem, integer was being compared with string for finding the issue. Thus, used toString to convert the id to string so that the type matches.